[otbn] Add OTBN encoding information to insns.yml
This patch defines a way to write "encoding schemes" in the YAML. This
is slightly inspired by the LLVM approach [1] but massively
simplified. I initially had something even simpler, without allowing
any hierarchy in the encoding schemes, but it was quite ugly to fill
in the instruction encodings and I was worried that it would make it
very difficult to change the encodings if we needed to later.

Instructions now have encodings defined, and we have code in the
Python parser that resolves the (hierarchical) scheme used by an
instruction into a flat one, then matches up the named operands with
the fields in the encoding scheme.

We finally check that the encoding is not ambiguous: that no bit
pattern matches more than one instruction.

Once this is all done, the documentation generator now spits out an
encoding table next to each instruction.

Note that there is currently no encoding for:

  - bn.mulqacc
  - bn.cmp
  - bn.cmpb

The current proposal for bn.mulqacc seems to have too many bits worth
of information to fit in the encoding space: I think we need to
revisit that design. And the proposed encoding that I was copying had
problems for bn.cmp and bn.cmpb (missing operands, and ambiguity with
bn.addi), so I've just dropped them for now.

[1] See e.g. llvm/lib/Target/RISCV/RISCVInstrFormatsC.td in the LLVM
    source tree.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>