[topgen] Remove explicit alert_module list from top_*.hjson
This list actually had code to populate it in topgen/merge.py's
amend_alert function, but this didn't ever run because it already has
a value. That value means that you need to remember to update it
whenever you add a new instance of an IP block.

But this seems a little silly: the code in amend_alert can be fixed to
do the right thing and the only difference between that and the manual
list is the exact ordering of the bits in the signal. Since we access
these bits through auto-generated named constants anyway, we can
dispense with the manual step entirely.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>