[otbn,dv] Ensure an error in opcode corrupting otbn_ctrl_redun_vseq
In this branch (if err_type is picked to be 2), the sequence tries to
corrupt the opcode in the bignum ALU. The existing code correctly
waited until the ALU had a valid operation, and then wrote a different
value to the opcode.

But the "valid" flag might be high when the operation was "None". If
that happened and we wrote another invalid operation (also ignored)
then the corruption wouldn't actually do anything. Oops!

With this patch, we wait for the operation_valid_i flag as before, but
also wait until the opcode specifies a valid operation. That way,
corrupting it should cause something to happen.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>