[otbn,dv] Rename and document a method in OtbnTraceEntry
The new name makes it a bit more obvious what is going on ("which guy
is the ISS entry?") and the documentation explains what the function
is supposed to be doing.

Note that the first argument is not of type OtbnIssTraceEntry. This is
for silly reasons: we're declaring the method on the base class. If we
define another subclass called "OtbnRtlTraceEntry" in the future, we
could move this method there and use a stricter type. We can give the
argument a more descriptive variable name, at least.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>