[sysrst_ctrl,doc] Slightly tweak intro text in theory of operation
One change is to move the "override" feature description to be closer
to the default pass-through behaviour. I think this makes the text a
little easier to follow.

I've also dropped the "This allows the security chip..." sentence. I
don't think the text is needed, and I'm not even completely sure who
the "main user" is.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>