[otbn] Make an assertion more local in otbn_start_stop_control.sv
The assertion was true inside our design, but you had to "go up a
level" to otbn_core to see why. I think it's probably cleaner if we
move the assertion to live somewhere it can be more local.

The reason that the new assertion is currently true:

  - The secure_wipe_req signal in otbn_core is an output from the
    controller (secure_wipe_req_o).

  - There, it is only turned on by the start_secure_wipe signal.

  - In turn, this signal is defined with

      assign start_secure_wipe = executing & (done_complete | err);

We can dig further to see why this assertion is equivalent to the one
that was there already:

  - The executing signal is defined by

      assign executing = (state_q == OtbnStateRun) ||
                         (state_q == OtbnStateStall);

  - There is an FSM that jumps between OtbnStateRun and
    OtbnStateStall, but we can only get into that pair of states when
    the start_i signal is sent to the controller.

  - That signal is an output from start_stop_control called
    controller_start_o. That signal is defined as:

      assign controller_start_o =
        ((state_q == OtbnStartStopStateUrndRefresh) & (urnd_reseed_ack_i | skip_reseed_q) &
          mubi4_test_false_strict(first_wipe_phase_done_q));

  - But this combination of signals only lasts for a single cycle,
    after which the start/stop controller will switch to
    OtbnStartStopStateRunning or OtbnStartStopStateLocked.

We don't really care if we get a secure_wipe_req_i signal when we're
in the locked state (since we'll ignore it anyway). So it would
suffice to show that

  (A) The controller won't start another new secure wipe request
      without starting a new run (which would require the start/stop
      controller to get back to OtbnStartStopStateRunning).

  (B) The start/stop controller won't leave OtbnStartStopStateRunning
      without a secure_wipe request.

I think that statement (A) is true because the main controller has a
reasonably close eye on the secure wipe state (through a
secure_wipe_ack signal).

Sadly (B) isn't quite true. The start/stop controller can also leave
the running state by receiving an escalation or RMA request. But both
of these cause a secure wipe and lock the controller.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>