[otbn,dv] Disable req fifo assertions in some FI tests
The assertions in tlul_adapter_sram are checking that the tl_o signal
that we send out is not unknown. The ones in OTBN itself are also
checking that the error bits that we expose are not unknown.

If we are corrupting the prim_count inside the prim_fifo_sync that
stores memory requests then we will have a bogus entry in the FIFO,
where the memory request is 'X. This turns into the tl_o signal,
causing the tlul_adapter_sram assertions to fail. The 'X response then
causes the OTBN assertions to fail and also causes failures in
prim_count (the loop counter), a prim_mubi4_sender (storing error
signals for wipes) and prim_alert_sender (sending the corrupted alert
signal).

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>