[otbn,rtl] Avoid locking on corrupt RMA signal
The RMA signal comes into the block with an rma_req_i port and is
encoded as lc_ctrl_pkg::lc_tx_t. If this happens to appear through a
CDC then the 4-bit signal might be corrupted for a cycle on the way
and the design is supposed to ignore it (until the value is strictly
equal to lc_ctrl_pkg::On.

For the ES chip, there was a bug here, where we would lock OTBN if
this happened. This bug was mostly fixed by commit 5228dfc, but that
isn't quite a complete fix. The problem is that rma_req_i gets sampled
and stored into rma_ack_d when the start/stop control block is in FSM
state OtbnStartStopSecureWipeComplete.

This is then clocked into rma_ack_q. The existing code correctly
squashed rma_ack_d back to MuBi4False if rma_ack_q was invalid. But it
also locked OTBN, which is different from the intended
behaviour (ignore invalid signals).

This commit is rather trivial and drops the code that was handling the
invalid signals.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>