[reggen] Move name, *sync_name and *sync_clk to RegBase
These are shared between Register and MultiRegister. Make the sharing
more explicit by moving the fields to the base class.

The changed code in reg_block.py used the underlying pseudo-register
to figure out a name and offset. Change it to use the information from
the MultiRegister directly: the offset was already there, and the name
is now. This should make the code a little more obvious.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>