[tlul_adapter_reg_tb,fpv] Properties for the A,D chan <-> reg iface
The assertions in the first section check that a TL message to read a
particular address will be correctly translated into a read
request (using re_o and addr_o) and that that re_o only happens if we
see a valid Get message.

This is quite fiddly to define, because there are a whole host of
conditions that must be satisfied by the Get message to avoid an
error (which causes the adapter to squash the read request).

They assertions in the first section also check that the TL interface
doesn't allow "overlapped" requests.

We then check that write requests are translated properly. This
includes checking that the we_o, addr_o, be_o and wdata_o signals
match the message.

We have covers for the two specific write opcodes (to check that e.g.
we don't only allow PutPartialData).

Next, we have some assertions about the D channel response. Although
TileLink allows a message sender to drop it at will, we don't want to
do that here, and so check this is true.

We also check that unknown opcodes get error responses and that error
responses have d_data == '1.

For responses to Get messages, we check that the data actually match
the stuff seen on rdata_i.

Finally, we have some basic checks about the register interface. In
particular, we check the busy_i and error_i signals have the expected
effects.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>