[alert,fpv] Tidy-ups in prim_alert_rxtx_assert_fpv.sv
This is inspired by trying to de-duplicate some of the four assertion
modules used for the "rxtx" tests but this commit is just supposed to
be tidying up the code in the file and documenting it more carefully.

We slightly simplify the AlertCheck0_A assertion: there's no need for
the different "disable iff" constants, since they were all implied by
the antecedent anyway.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>